Drop archived automated_deployment source tree into /home/vagrant.
The archive must exist in the vagrant/control dir so that it can
be shared with the guest.  Since the two git projects are separate
at the moment, this is just an easy way to get by.

To set this up, assume that the automation repo is in a directory
next to this repo:

  tar cfz automated_deployment.tgz automated_deployment && \
      mv automated_deployment.tgz ~/vagrant/control